refactor: Move string_functions tests to dedicated test file
Migrate 101 string function tests to tests/string_functions.rs
to improve code organization and maintainability
