+++ src/vm/tests/mod.rs
@@ -2,3 +2,4 @@ pub mod helpers;
 mod basic;
 mod math_variadic;
 mod math_errors;
+mod string_functions;
